// src/components/atoms/Button/type.ts
export enum ButtonVariants {
  OUTLINE = "OUTLINE",
  LINK = "LINK",
}
export enum ButtonColorSchemes {
  PRIMARY = "PRIMARY",
  BLACK = "BLACK",
  GRAY = "GRAY",
  WHITE = "WHITE",
}
export enum ButtonSizes {
  TINY = "TINY",
  SMALL = "SMALL",
  MEDIUM = "MEDIUM",
  LARGE = "LARGE",
  XLAGRE = "XLAGRE",
  XXLARGE = "XXLARGE",
}
export enum ButtonShapes {
  SQUARE = "SQUARE",
  ROUNDED = "ROUNDED",
  CIRCLE = "CIRCLE",
}